Added Naming of the Type | Added check for Quitting
Added Naming of the Type. Before it was just naming it a blank string. Now it names it whatever the calling type name is.

Added check for Quitting. I had this before but apparently removed it for some reason. This keeps GameObjects from being created after Unity has already started shutting down (or coming out of play)